requirements.txt
langchain>=0.1.0
langchain-core>=0.1.16
langchain-community>=0.0.19
langchain-deepseek>=0.0.1
python-dotenv>=1.0.0
weaviate-client>=3.25.0
langchain-weaviate>=0.0.1
langchain-text-splitters>=0.0.1
pypdf>=3.15.1
unstructured>=0.10.30
tabulate>=0.9.0
pdf2image>=1.16.3
pillow>=10.0.1
flask>=2.0.0
werkzeug>=2.0.0
flask-cors>=3.0.10 